+++ jbpm/jbpm-flow/src/main/java/org/jbpm/process/instance/AbstractProcessRuntime.java
@@ -22,13 +22,23 @@ import java.util.IdentityHashMap;
 import java.util.List;
 import java.util.Map;
 
+import org.drools.base.time.TimeUtils;
+import org.drools.core.common.InternalKnowledgeRuntime;
+import org.jbpm.process.core.timer.DateTimeUtils;
+import org.jbpm.process.core.timer.Timer;
 import org.jbpm.process.instance.event.KogitoProcessEventListenerAdapter;
 import org.jbpm.process.instance.event.KogitoProcessEventSupportImpl;
 import org.kie.api.event.process.ProcessEventListener;
 import org.kie.kogito.Application;
+import org.kie.kogito.calendar.BusinessCalendar;
 import org.kie.kogito.internal.process.event.KogitoProcessEventListener;
 import org.kie.kogito.internal.process.event.KogitoProcessEventSupport;
 import org.kie.kogito.internal.process.runtime.KogitoProcessRuntime;
+import org.kie.kogito.jobs.DurationExpirationTime;
+import org.kie.kogito.jobs.ExactExpirationTime;
+import org.kie.kogito.jobs.ExpirationTime;
+
+import static org.jbpm.process.core.constants.CalendarConstants.BUSINESS_CALENDAR_ENVIRONMENT_KEY;
 
 public abstract class AbstractProcessRuntime implements InternalProcessRuntime {
 
@@ -84,5 +94,60 @@ public abstract class AbstractProcessRuntime implements InternalProcessRuntime {
             return ((KogitoProcessEventListener) processEventListener);
         }
         return listenersMap.remove(processEventListener);
+    }
+
+    protected ExpirationTime createTimerInstance(Timer timer, InternalKnowledgeRuntime kruntime) {
+
+        if (kruntime != null && kruntime.getEnvironment().get(BUSINESS_CALENDAR_ENVIRONMENT_KEY) != null) {
+            BusinessCalendar businessCalendar = (BusinessCalendar) kruntime.getEnvironment().get(BUSINESS_CALENDAR_ENVIRONMENT_KEY);
+
+            long delay = businessCalendar.calculateBusinessTimeAsDuration(timer.getDelay());
+
+            if (timer.getPeriod() == null) {
+                return DurationExpirationTime.repeat(delay);
+            } else {
+                long period = businessCalendar.calculateBusinessTimeAsDuration(timer.getPeriod());
+
+                return DurationExpirationTime.repeat(delay, period);
+            }
+        } else {
+            return configureTimerInstance(timer);
+        }
+    }
+
+    private ExpirationTime configureTimerInstance(Timer timer) {
+        long duration = -1;
+        switch (timer.getTimeType()) {
+            case Timer.TIME_CYCLE:
+                // when using ISO date/time period is not set
+                long[] repeatValues = DateTimeUtils.parseRepeatableDateTime(timer.getDelay());
+                if (repeatValues.length == 3) {
+                    return DurationExpirationTime.repeat(repeatValues[1], repeatValues[2]);
+                } else {
+                    long delay = repeatValues[0];
+                    long period = -1;
+                    try {
+                        period = TimeUtils.parseTimeString(timer.getPeriod());
+
+                    } catch (RuntimeException e) {
+                        period = repeatValues[0];
+                    }
+
+                    return DurationExpirationTime.repeat(delay, period);
+                }
+
+            case Timer.TIME_DURATION:
+
+                duration = DateTimeUtils.parseDuration(timer.getDelay());
+                return DurationExpirationTime.repeat(duration);
+
+            case Timer.TIME_DATE:
+
+                return ExactExpirationTime.of(timer.getDate());
+
+            default:
+                throw new UnsupportedOperationException("Not supported timer definition");
+        }
+
     }
 }
