+++ jbpm/jbpm-flow/src/main/java/org/jbpm/process/instance/LightProcessRuntime.java
@@ -24,15 +24,12 @@ import java.util.List;
 import java.util.Map;
 
 import org.drools.base.definitions.rule.impl.RuleImpl;
-import org.drools.base.time.TimeUtils;
 import org.drools.core.common.InternalKnowledgeRuntime;
 import org.drools.core.common.ReteEvaluator;
 import org.drools.core.common.WorkingMemoryAction;
 import org.drools.core.phreak.PropagationEntry;
 import org.jbpm.process.core.event.EventFilter;
 import org.jbpm.process.core.event.EventTypeFilter;
-import org.jbpm.process.core.timer.DateTimeUtils;
-import org.jbpm.process.core.timer.Timer;
 import org.jbpm.ruleflow.core.RuleFlowProcess;
 import org.jbpm.workflow.core.impl.DataAssociation;
 import org.jbpm.workflow.core.impl.NodeIoHelper;
@@ -51,12 +48,8 @@ import org.kie.api.runtime.rule.AgendaFilter;
 import org.kie.internal.process.CorrelationKey;
 import org.kie.internal.runtime.StatefulKnowledgeSession;
 import org.kie.kogito.Application;
-import org.kie.kogito.calendar.BusinessCalendar;
 import org.kie.kogito.internal.process.runtime.KogitoProcessInstance;
 import org.kie.kogito.internal.process.workitem.KogitoWorkItemManager;
-import org.kie.kogito.jobs.DurationExpirationTime;
-import org.kie.kogito.jobs.ExactExpirationTime;
-import org.kie.kogito.jobs.ExpirationTime;
 import org.kie.kogito.jobs.JobsService;
 import org.kie.kogito.jobs.ProcessJobDescription;
 import org.kie.kogito.process.Processes;
@@ -64,7 +57,6 @@ import org.kie.kogito.services.jobs.impl.InMemoryJobService;
 import org.kie.kogito.signal.SignalManager;
 import org.kie.kogito.uow.UnitOfWorkManager;
 
-import static org.jbpm.process.core.constants.CalendarConstants.BUSINESS_CALENDAR_ENVIRONMENT_KEY;
 import static org.jbpm.ruleflow.core.Metadata.TRIGGER_MAPPING_INPUT;
 
 public class LightProcessRuntime extends AbstractProcessRuntime {
@@ -423,61 +415,6 @@ public class LightProcessRuntime extends AbstractProcessRuntime {
     public boolean isActive() {
         // originally: kruntime.getEnvironment().get("Active")
         return runtimeContext.isActive();
-    }
-
-    protected ExpirationTime createTimerInstance(Timer timer, InternalKnowledgeRuntime kruntime) {
-
-        if (kruntime != null && kruntime.getEnvironment().get(BUSINESS_CALENDAR_ENVIRONMENT_KEY) != null) {
-            BusinessCalendar businessCalendar = (BusinessCalendar) kruntime.getEnvironment().get(BUSINESS_CALENDAR_ENVIRONMENT_KEY);
-
-            long delay = businessCalendar.calculateBusinessTimeAsDuration(timer.getDelay());
-
-            if (timer.getPeriod() == null) {
-                return DurationExpirationTime.repeat(delay);
-            } else {
-                long period = businessCalendar.calculateBusinessTimeAsDuration(timer.getPeriod());
-
-                return DurationExpirationTime.repeat(delay, period);
-            }
-        } else {
-            return configureTimerInstance(timer);
-        }
-    }
-
-    private ExpirationTime configureTimerInstance(Timer timer) {
-        long duration = -1;
-        switch (timer.getTimeType()) {
-            case Timer.TIME_CYCLE:
-                // when using ISO date/time period is not set
-                long[] repeatValues = DateTimeUtils.parseRepeatableDateTime(timer.getDelay());
-                if (repeatValues.length == 3) {
-                    return DurationExpirationTime.repeat(repeatValues[1], repeatValues[2]);
-                } else {
-                    long delay = repeatValues[0];
-                    long period = -1;
-                    try {
-                        period = TimeUtils.parseTimeString(timer.getPeriod());
-
-                    } catch (RuntimeException e) {
-                        period = repeatValues[0];
-                    }
-
-                    return DurationExpirationTime.repeat(delay, period);
-                }
-
-            case Timer.TIME_DURATION:
-
-                duration = DateTimeUtils.parseDuration(timer.getDelay());
-                return DurationExpirationTime.repeat(duration);
-
-            case Timer.TIME_DATE:
-
-                return ExactExpirationTime.of(timer.getDate());
-
-            default:
-                throw new UnsupportedOperationException("Not supported timer definition");
-        }
-
     }
 
     public class SignalManagerSignalAction extends PropagationEntry.AbstractPropagationEntry implements WorkingMemoryAction {
