+++ jbpm/jbpm-flow/src/main/java/org/jbpm/process/instance/ProcessRuntimeImpl.java
@@ -24,7 +24,6 @@ import java.util.List;
 import java.util.Map;
 
 import org.drools.base.definitions.rule.impl.RuleImpl;
-import org.drools.base.time.TimeUtils;
 import org.drools.core.common.InternalKnowledgeRuntime;
 import org.drools.core.common.InternalWorkingMemory;
 import org.drools.core.common.ReteEvaluator;
@@ -35,8 +34,6 @@ import org.drools.core.time.impl.CommandServiceTimerJobFactoryManager;
 import org.drools.core.time.impl.ThreadSafeTrackableTimeJobFactoryManager;
 import org.jbpm.process.core.event.EventFilter;
 import org.jbpm.process.core.event.EventTypeFilter;
-import org.jbpm.process.core.timer.DateTimeUtils;
-import org.jbpm.process.core.timer.Timer;
 import org.jbpm.process.instance.event.DefaultSignalManagerFactory;
 import org.jbpm.process.instance.event.KogitoProcessEventSupportImpl;
 import org.jbpm.process.instance.impl.DefaultProcessInstanceManagerFactory;
@@ -63,12 +60,8 @@ import org.kie.internal.command.RegistryContext;
 import org.kie.internal.process.CorrelationKey;
 import org.kie.internal.runtime.StatefulKnowledgeSession;
 import org.kie.kogito.Application;
-import org.kie.kogito.calendar.BusinessCalendar;
 import org.kie.kogito.internal.process.runtime.KogitoProcessInstance;
 import org.kie.kogito.internal.process.runtime.KogitoProcessRuntime;
-import org.kie.kogito.jobs.DurationExpirationTime;
-import org.kie.kogito.jobs.ExactExpirationTime;
-import org.kie.kogito.jobs.ExpirationTime;
 import org.kie.kogito.jobs.JobsService;
 import org.kie.kogito.jobs.ProcessJobDescription;
 import org.kie.kogito.services.identity.NoOpIdentityProvider;
@@ -78,7 +71,6 @@ import org.kie.kogito.services.uow.DefaultUnitOfWorkManager;
 import org.kie.kogito.signal.SignalManager;
 import org.kie.kogito.uow.UnitOfWorkManager;
 
-import static org.jbpm.process.core.constants.CalendarConstants.BUSINESS_CALENDAR_ENVIRONMENT_KEY;
 import static org.jbpm.ruleflow.core.Metadata.TRIGGER_MAPPING_INPUT;
 
 public class ProcessRuntimeImpl extends AbstractProcessRuntime {
@@ -407,60 +399,6 @@ public class ProcessRuntimeImpl extends AbstractProcessRuntime {
         }
 
         return active.booleanValue();
-    }
-
-    protected ExpirationTime createTimerInstance(Timer timer, InternalKnowledgeRuntime kruntime) {
-        if (kruntime != null && kruntime.getEnvironment().get(BUSINESS_CALENDAR_ENVIRONMENT_KEY) != null) {
-            BusinessCalendar businessCalendar = (BusinessCalendar) kruntime.getEnvironment().get(BUSINESS_CALENDAR_ENVIRONMENT_KEY);
-
-            long delay = businessCalendar.calculateBusinessTimeAsDuration(timer.getDelay());
-
-            if (timer.getPeriod() == null) {
-                return DurationExpirationTime.repeat(delay);
-            } else {
-                long period = businessCalendar.calculateBusinessTimeAsDuration(timer.getPeriod());
-
-                return DurationExpirationTime.repeat(delay, period);
-            }
-        } else {
-            return configureTimerInstance(timer);
-        }
-    }
-
-    private ExpirationTime configureTimerInstance(Timer timer) {
-        long duration = -1;
-        switch (timer.getTimeType()) {
-            case Timer.TIME_CYCLE:
-                // when using ISO date/time period is not set
-                long[] repeatValues = DateTimeUtils.parseRepeatableDateTime(timer.getDelay());
-                if (repeatValues.length == 3) {
-                    int parsedReapedCount = (int) repeatValues[0];
-
-                    return DurationExpirationTime.repeat(repeatValues[1], repeatValues[2], parsedReapedCount);
-                } else {
-                    long delay = repeatValues[0];
-                    long period = -1;
-                    try {
-                        period = TimeUtils.parseTimeString(timer.getPeriod());
-                    } catch (RuntimeException e) {
-                        period = repeatValues[0];
-                    }
-
-                    return DurationExpirationTime.repeat(delay, period);
-                }
-
-            case Timer.TIME_DURATION:
-
-                duration = DateTimeUtils.parseDuration(timer.getDelay());
-                return DurationExpirationTime.after(duration);
-
-            case Timer.TIME_DATE:
-
-                return ExactExpirationTime.of(timer.getDate());
-
-            default:
-                throw new UnsupportedOperationException("Not supported timer definition");
-        }
     }
 
     @Override
